+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The supported Python versions are now 3.11, 3.11, 3.12, 3.13, and 3.14.
 - Dev: Added some unit tests for \$(randomchoice:...). (#2839)
 - Dev: Added a `test.sh` script that errors if we use something deprecated. (#2855)
 - Dev: Added unit test for `utils.now`. (#2856)
+- Dev: Added unit test for `config.get_boolean`. (#2860)
 - Dev: Remove dependency on the `cgi` package. (#2827)
 - Dev: Requirement files are now sourced from pyproject.toml. `uv` is recommended to install dependencies. (#2826, #2841)
 - Dev: Use our own fork of `pytimeparse` to ensure we don't error to fix Python3.14 deprecation warnings. (#2857)